Reorganized files, updated file references, and added import comments
