+++ screensaver-display/src/pages/UserHome.js
@@ -1,3 +1,4 @@
+// dependencies
 import React from 'react';
 
 const UserHome = () => {
