+++ screensaver-display/src/pages/UserRedditSearch.js
@@ -1,14 +1,20 @@
+// dependencies
 import React, { useState } from 'react';
 import Images from './PostImages';
-import styles from './redditsearch.module.css';
-import { SearchBar } from '../components/SearchBar.js';
-import { Button } from '../components/Button.js';
+
+// UI dependencies
+import { Button } from '../ui-components/Button.js';
 import { LazyLoadImage } from 'react-lazy-load-image-component';
+
+// styles
+import styles from './styles/userredditsearch.module.css';
+import './styles/RedditSearchbar.css';
 import 'react-lazy-load-image-component/src/effects/blur.css';
 
-export default function UserRedditSearch() {
+const UserRedditSearch = () => {
 	const [ selectedImg, setSelectedImg ] = useState(Images[0]);
-	const [ query, setQuery ] = React.useState('');
+	const [ query, setQuery ] = useState('');
+	const [ subreddits, setSubreddits] = useState([]);
 	
 	return (
 		<div className={styles['container']}>
@@ -17,41 +23,52 @@ export default function UserRedditSearch() {
 				<Button type="enable" variant="small">
 					Reddit Mode
 				</Button>
-				<SearchBar query={query} setQuery={setQuery} />
+
+
 			</div>
 			<div className={styles['imgContainer']}>
 				{Images.map((img, index) => (
 					<LazyLoadImage
-					style={{ border: selectedImg === img ? '4px solid green' : '' }}
 					key={index}
 					src={img}
 					alt="dog"
-					onClick={() => setSelectedImg(img)}
 					effect="blur"
 					/>
 				))}
 			</div>
 		</div>
 		
-		// <div className={styles['UserPosts']}>
-		// 	<div className={styles['container']}>
-		// 		<SearchBar query={query} setQuery={setQuery} />
-		// 		<div className={styles['imgContainer']}>
-		// 			{Images.map((img, index) => (
-		// 				<LazyLoadImage
-		// 					style={{ border: selectedImg === img ? '4px solid green' : '' }}
-		// 					key={index}
-		// 					src={img}
-		// 					alt="dog"
-		// 					onClick={() => setSelectedImg(img)}
-		// 					effect="blur"
-		// 				/>
-		// 			))}
-		// 		</div>
-		// 		<Button type="enable" variant="small">
-		// 			Reddit Mode
-		// 		</Button>
-		// 	</div>
-		// </div>
+
 	);
 }
+
+
+
+const Results = (props) => {
+
+    const returnList = () => {
+        const defaultIcon = 'https://styles.redditmedia.com/t5_6/styles/communityIcon_a8uzjit9bwr21.png?width=256&s=d28ea66f16da5a6c2ccae0d069cc4d42322d69a9';
+        return props.posts.map((post, index) => {
+            return (
+                <div className='result' key={index}>
+                    <div className='subredditThumbnail'>
+                        <img src={post.data.header_img || defaultIcon} alt={`r/${post.data.display_name}`}/>
+                    </div>
+                    <p className='subredditTitle'>r/{post.data.display_name}</p>
+                </div>
+            );
+        })
+    }
+    
+    return (
+        <>
+        <div className='resultsBox'>
+            {returnList()}
+        </div>
+
+        </>
+    );
+
+};
+
+export default UserRedditSearch;
