+++ screensaver-display/src/pages/styles/RedditSearchbar.css
@@ -1,0 +1,49 @@
+
+
+.currentSubreddit {
+    width: 50%;
+    margin: auto;
+}
+
+h1 {
+    text-align: center;
+}
+
+.resultsBox {
+    width: 75%;
+    margin: auto;
+}
+
+.result {
+    padding: 5px;
+    margin-top: 15px;
+    background-color: lightgray;
+    border-radius: 5px;
+    height: 60px;
+}
+
+.subredditThumbnail {
+    margin: 0 auto;
+    display: inline-block;
+    text-align: center;
+    height: 10px;
+    width: 10px;
+    overflow: hidden; 
+    border: 1px solid black;
+    border-radius: 50%;
+}
+
+.subredditThumbnail > img {
+    overflow: hidden;
+    display: block;
+    margin: auto;
+    height: 100%;
+    width: auto;
+}
+
+.subredditTitle {
+    display: inline;
+    margin: 5px;
+    font-weight: 300;
+    font-size: 10px;
+}
