+++ screensaver-display/src/pages/styles/slideshow.module.css
@@ -1,4 +1,4 @@
-@import "../styles/main.css";
+@import "../../styles/main.css";
 
 .center {
   display: block;
