+++ screensaver-display/src/pages/styles/userredditsearch.module.css
@@ -18,18 +18,13 @@
 	height: 100%;
 }
 
-.container .selected {
-	width: 500px;
-	height: 500px;
-	margin: 0 auto;
-	border: 4px solid green;
-}
+
 
 .searchContainer {
 	width: 350px;
 	margin-right: 20px;
-	border-right: 1px solid #888;
-	box-shadow: 5px 0 5px #ccc;
+	border-right: 1px solid #bbb;
+	box-shadow: 3px 0 10px #ccc;
 	background-color: #eee;
 }
 
@@ -45,8 +40,8 @@
 
 .imgContainer img {
 	cursor: pointer;
-	width: 100px !important;
-	height: 100px !important;
+	width: 200px !important;
+	height: 200px !important;
 	border-radius: 6px;
 	box-shadow: 3px 3px 5px 6px #ccc;
 	margin: 0.5rem 1rem;
