+++ screensaver-display/src/ui-components/ThumbnailImage.js
@@ -1,5 +1,10 @@
+// dependencies
 import React from 'react';
+
+// UI dependencies
 import DEFAULT_IMAGE from '../styles/ExampleImage2.jpg';
+
+// styles
 import styles from './thumbnailimage.module.css';
 
 export const ThumbnailImage = () => {
