+++ screensaver-display/src/ui-components/styles/button.module.css
@@ -1,4 +1,4 @@
-@import "../styles/main.css";
+@import "../../styles/main.css";
 
 .btn-primary {
 	margin: 34px auto;
