+++ screensaver-display/src/ui-components/styles/sidenavbar.module.css
@@ -1,4 +1,4 @@
-@import "../styles/main.css";
+@import "../../styles/main.css";
 
 .sidenavbar {
 	/* background-color: var(--color-grey-light); */
