+++ screensaver-display/src/ui-components/styles/thumbnailimage.module.css
