+++ screensaver-display/src/ui-components/styles/typography.module.css
@@ -1,4 +1,4 @@
-@import "../styles/main.css";
+@import "../../styles/main.css";
 
 .paragraph {
 	font-family: var(--font-Montserrat);
